refactor(stateappevents): split projector source by responsibility
stateappevents_projector.go did four jobs in 373 lines. Split into focused
files (package unchanged; functions share the logState struct, not control flow,
so this is a clean file split, not a leaky abstraction):

  stateappevents_projector.go  lifecycle: logState, registerLog, startProjector,
                               shuttingDown, logProjection
  stateappevents_coldstart.go  cold-start snapshot seeding: coldStartFrom,
                               seedFromSnapshot, haltUnbridgeable, haltErasure
  stateappevents_gap.go        gap classification: applyRecord, classifyGap, gapClass
  stateappevents_fold.go       the fold: projectRecord

Tests untouched this commit (redistribution follows); go vet + go test -race green.
