+++ stateappevents_projector.go
@@ -2,9 +2,6 @@ package via
 
 import (
 	"context"
-	"encoding/json"
-	"errors"
-	"reflect"
 	"sync"
 	"time"
 )
@@ -102,64 +99,6 @@ func (a *App) startProjector(key string, ls *logState) {
 	}()
 }
 
-// coldStartFrom seeds ls from a durable snapshot (so the projector replays only
-// the tail) and returns the offset to subscribe from. A missing / stale-codec /
-// undecodable snapshot returns 0 (re-fold from genesis) — the snapshot is a
-// disposable cache, never required for correctness. Returns the snapshot's
-// covered offset only when a bridging seed (matching codec or seeded migration)
-// was installed; otherwise a halt is latched on ls and folding stops.
-func (a *App) coldStartFrom(ls *logState, key string) Offset {
-	data, rev, ok, _ := a.backplane.LoadSnapshot(context.Background(), snapKey(key))
-	if !ok {
-		return 0
-	}
-	var cp checkpoint
-	if json.Unmarshal(data, &cp) != nil {
-		return 0
-	}
-	// A snapshot folded BEFORE a crypto-shred erasure may hold shredded PII; if
-	// the authoritative generation has advanced past it, it is invalid.
-	erasureStale := cp.ErasureGen < a.loadErasureGen()
-	switch {
-	case erasureStale && cp.Compacted:
-		// Durable-genesis snapshot invalidated by erasure: the event prefix is
-		// gone, so re-folding from 0 would silently TRUNCATE to the surviving
-		// tail — dropping SURVIVING subjects' data, not only the erased one. Halt
-		// (roll-forward-only); compaction + erasure of one key needs snapshot
-		// re-encryption (documented follow-up).
-		a.haltErasure(ls, key)
-		return 0
-	case erasureStale:
-		// Uncompacted + stale → re-fold from genesis; the erased subject's
-		// now-undecryptable events drop on the way.
-		return 0
-	case cp.CodecHash == ls.codecHash:
-		// Codec matches → seed straight from the snapshot.
-		if v, err := ls.decodeSnap(cp.V); err == nil {
-			a.seedFromSnapshot(ls, v, cp, rev)
-			return cp.CoveredOffset
-		}
-		return 0
-	case !cp.Compacted:
-		// Uncompacted mismatch → the snapshot is a pure disposable cache; re-fold
-		// from genesis. Evolving V is free.
-		return 0
-	default:
-		// Compacted + mismatch → durable genesis: the event prefix is gone, so we
-		// MUST NOT discard (that would silently truncate to the surviving tail).
-		// Run the registered seeded migration; on a missing or failing migration
-		// HALT the projector (roll-forward-only), never truncate (T2-GO-4).
-		if migrate, found := lookupSnapMigration(cp.CodecHash); found {
-			if v, err := migrate.decode(cp.V); err == nil {
-				a.seedFromSnapshot(ls, v, cp, rev)
-				return cp.CoveredOffset
-			}
-		}
-		a.haltUnbridgeable(ls, key)
-		return cp.CoveredOffset
-	}
-}
-
 // shuttingDown reports whether Shutdown has begun (backplaneDone closed), so a
 // projector/consumer can tell a graceful stop from a transient disconnect.
 func (a *App) shuttingDown() bool {
@@ -167,193 +106,6 @@ func (a *App) shuttingDown() bool {
 	case <-a.backplaneDone:
 		return true
 	default:
-		return false
-	}
-}
-
-// seedFromSnapshot installs a cold-start seed (from a matching snapshot or a
-// seeded migration) as the projection baseline, resuming the tail at the
-// snapshot's covered offset.
-func (a *App) seedFromSnapshot(ls *logState, v any, cp checkpoint, rev Rev) {
-	ls.mu.Lock()
-	ls.projection = v
-	ls.cursor = cp.CoveredOffset
-	ls.epoch = cp.Epoch
-	ls.epochSeen = true
-	ls.snapRev = rev
-	ls.mu.Unlock()
-}
-
-// haltUnbridgeable freezes a compacted key whose durable-genesis snapshot cannot
-// be bridged to the current fold (no / failing migration). Roll-forward-only:
-// the projector folds nothing further (projectRecord short-circuits on halted),
-// so the value never silently truncates to the surviving tail.
-func (a *App) haltUnbridgeable(ls *logState, key string) {
-	ls.mu.Lock()
-	ls.halted = true
-	ls.mu.Unlock()
-	a.metricsOrNoop().Counter("via.snapshot.unbridgeable", "key", key)
-}
-
-// haltErasure freezes a key whose COMPACTED snapshot was invalidated by a
-// crypto-shred erasure: its event prefix is gone, so neither seeding the
-// (PII-bearing, stale) snapshot nor re-folding from a truncated log is correct.
-// Roll-forward-only until the operator re-encrypts the snapshot (the documented
-// compaction+erasure follow-up), so surviving subjects' data is never silently
-// truncated.
-func (a *App) haltErasure(ls *logState, key string) {
-	ls.mu.Lock()
-	ls.halted = true
-	ls.mu.Unlock()
-	a.metricsOrNoop().Counter("via.snapshot.erasure_halt", "key", key)
-}
-
-// applyRecord handles one delivered record. It first guards against COMPACTION
-// OVERTAKE: if this projector has folded up to cursor C but the next record's
-// offset is beyond C+1, the log dropped records C+1..(offset-1) that this pod
-// never folded (a faster peer compacted the shared log below this pod's
-// position). Folding the post-gap record onto the stale projection would
-// silently and permanently diverge this pod from its peers — the cross-pod
-// truncation bug. Instead re-seed from the durable snapshot (which covers the
-// gap) and let the fold continue; if no snapshot can recover the gap, HALT
-// rather than diverge. Then delegate to projectRecord. Runs on the single
-// projector goroutine, so the read-then-reseed is not racing another folder.
-func (a *App) applyRecord(ls *logState, key string, rec Record) bool {
-	ls.mu.RLock()
-	cur, halted, gapsBenign := ls.cursor, ls.halted, ls.gapsBenign
-	ls.mu.RUnlock()
-	// rec.Offset > cur+1 is a gap: offsets cur+1..rec.Offset-1 were not delivered.
-	// That can mean a compacted prefix (real data loss) OR simply a backend whose
-	// per-key offsets are not contiguous — e.g. a JetStream stream sequenced
-	// GLOBALLY across subjects, where each key's records carry the shared stream
-	// sequence and skip the numbers other keys took. classifyGap tells them apart;
-	// once a backend proves non-contiguous we latch (gapsBenign) and stop checking,
-	// so a busy shared stream does not pay a snapshot read per record.
-	if !halted && !gapsBenign && rec.Offset > cur+1 {
-		switch a.classifyGap(ls, key, cur, rec.Offset-1) {
-		case gapReseeded:
-			a.metricsOrNoop().Counter("via.events.compaction_reseed", "key", key)
-		case gapHalt:
-			ls.mu.Lock()
-			ls.halted = true
-			ls.mu.Unlock()
-			a.metricsOrNoop().Counter("via.events.compaction_gap_halt", "key", key)
-			return false
-		case gapBenign:
-			// Not a lost prefix — fold the record and latch so subsequent
-			// non-contiguous records skip the gap check (and its snapshot read).
-			ls.mu.Lock()
-			ls.gapsBenign = true
-			ls.mu.Unlock()
-		}
-	}
-	return a.projectRecord(ls, key, rec)
-}
-
-// gapClass is how applyRecord interprets an offset gap.
-type gapClass int
-
-const (
-	gapBenign   gapClass = iota // not a lost prefix → fold the record
-	gapReseeded                 // recovered from a bridging snapshot
-	gapHalt                     // a compacted prefix we cannot bridge → freeze (never truncate)
-)
-
-// classifyGap decides what a gap (rec.Offset > cursor+1) means. A gap is a lost
-// prefix ONLY if compaction discarded it, and compaction always writes a
-// Compacted snapshot FIRST (snapshot-FIRST). So:
-//   - no snapshot, or an unreadable one          → benign (no compaction ran)
-//   - a snapshot that bridges the gap            → reseed from it (projection=V,
-//     cursor=CoveredOffset), recovering the dropped prefix instead of skipping it
-//   - a Compacted snapshot that cannot bridge    → halt, never truncate to the tail
-//   - an uncompacted snapshot that cannot bridge → benign (it is a disposable
-//     cache; the gap is just non-contiguous offsets, nothing was lost)
-//
-// The backplane read is OUTSIDE ls.mu; seedFromSnapshot takes the lock.
-func (a *App) classifyGap(ls *logState, key string, cur, needCovered Offset) gapClass {
-	data, rev, ok, _ := a.backplane.LoadSnapshot(context.Background(), snapKey(key))
-	if !ok {
-		return gapBenign
-	}
-	var cp checkpoint
-	if json.Unmarshal(data, &cp) != nil {
-		return gapBenign
-	}
-	// Recover only if the snapshot BRIDGES the gap (covers up to the record before
-	// rec) and moves us FORWARD — never seed behind our current position.
-	if cp.CodecHash == ls.codecHash && cp.CoveredOffset >= needCovered && cp.CoveredOffset > cur {
-		if v, err := ls.decodeSnap(cp.V); err == nil {
-			a.seedFromSnapshot(ls, v, cp, rev)
-			return gapReseeded
-		}
-	}
-	if cp.Compacted {
-		return gapHalt
-	}
-	return gapBenign
-}
-
-// projectRecord folds one delivered Record into the cached projection under the
-// key's lock, returning whether the projection advanced (and a re-render is due).
-// It is the single fold path — shared by the live projector loop. broadcastRender
-// is the caller's job, OUTSIDE the lock.
-func (a *App) projectRecord(ls *logState, key string, rec Record) (advanced bool) {
-	ls.mu.Lock()
-	defer ls.mu.Unlock()
-	if ls.halted {
-		return false
-	}
-	// Epoch / offset-space-reset detection (T1-SRE-3). The first record sets the
-	// baseline; a later record on a DIFFERENT epoch means the stream was
-	// recreated/trimmed/restored and its offsets restarted — a bare offset
-	// high-water-mark would skip every new record, so re-snapshot from genesis.
-	if !ls.epochSeen {
-		ls.epoch = rec.Epoch
-		ls.epochSeen = true
-	} else if rec.Epoch != ls.epoch {
-		ls.projection = ls.seed
-		ls.cursor = 0
-		ls.epoch = rec.Epoch
-		a.metricsOrNoop().Counter("via.events.epoch_reset", "key", key)
-	}
-	if rec.Offset <= ls.cursor {
-		return false
-	}
-	next, ferr := ls.foldBytes(ls.projection, rec.Data)
-	switch {
-	case ferr == nil:
-		if a.cfg.foldVerify && !ls.diverged {
-			// Re-fold the SAME record from the SAME accumulator: a deterministic
-			// reducer must produce an identical result. A mismatch (or a verify
-			// error) proves impurity — flag the key so it is never compacted, so
-			// the bad projection can't be crystallized into durable genesis.
-			if next2, verr := ls.foldBytes(ls.projection, rec.Data); verr != nil || !reflect.DeepEqual(next, next2) {
-				ls.diverged = true
-				a.metricsOrNoop().Counter("via.fold.divergence", "key", key)
-			}
-		}
-		ls.projection = next
-		ls.cursor = rec.Offset
-		ls.foldsSinceSnap++
-		return true
-	case errors.Is(ferr, ErrForwardIncompatible):
-		// A newer binary wrote this record. FREEZE this key — do NOT advance the
-		// cursor, so a roll-forward redeploy resumes here.
-		ls.halted = true
-		a.metricsOrNoop().Counter("via.events.forward_incompatible", "key", key)
-		return false
-	case errors.Is(ferr, ErrErased):
-		// The data subject was erased (crypto-shred): the payload is permanently
-		// unreadable BY DESIGN. Skip + advance, exactly like a poison record, but
-		// record it as an intentional erasure, not corruption.
-		ls.cursor = rec.Offset
-		a.metricsOrNoop().Counter("via.events.erased", "key", key)
-		return false
-	default:
-		// Poison / undecodable: skip it (advance past so it is not retried
-		// forever), never wedging the key for any pod.
-		ls.cursor = rec.Offset
-		a.metricsOrNoop().Counter("via.events.undecodable", "key", key)
 		return false
 	}
 }
